Fix build errors with recent boost libraries (>= 1.67)
With https://github.com/boostorg/date_time/commit/f9f2aaf5216c
we need to update the usage of duration.
Reference: https://github.com/Icinga/icinga2/pull/6230

+++ rostime/include/ros/impl/duration.h
@@ -184,9 +184,9 @@ namespace ros {
   {
     namespace bt = boost::posix_time;
 #if defined(BOOST_DATE_TIME_HAS_NANOSECONDS)
-    return bt::seconds(sec) + bt::nanoseconds(nsec);
+    return bt::seconds(sec) + bt::nanoseconds(long(nsec));
 #else
-    return bt::seconds(sec) + bt::microseconds(nsec/1000);
+    return bt::seconds(sec) + bt::microseconds(long(nsec/1000));
 #endif
   }
 }
